Create DDLmk2.sol
the system requires a user to deposit un-withdrawable collateral into the contract, so that they do not just deposit collateral and withdraw it. this is ONLY usable to pay debt. you need to put up a selected amount of your collateral, and it let's you see the amount of lends paid back on an account.